requirements.txt
Flask==2.3.2
flask-socketio>=5.1.1
eventlet==0.33.3
redis==4.6.0
flask_session==0.5.0
greenlet==1.1.3.post0
requests==2.28.1
python-socketio[client]==5.8.0
pyyaml>=6.0
aiohttp>=3.8.3
redis==4.6.0
python-dotenv>=0.21.0
python-arango>=7.6.2
pycryptodome>=3.18.0
urllib3[secure]==1.26.12
numpy>=1.23.1